Style for kinect integration
