+++ catkin_home/src/arm_vision/include/IncludersCloud/IncludersCloud.h
@@ -1,14 +1,16 @@
+#ifndef INCLUDERSCLOUD_H_
+#define INCLUDERSCLOUD_H_
+
+//TODO cleanup includerscloud
+
 #include <ros/ros.h>
 #include <sensor_msgs/PointCloud2.h>
 
 #include <iostream>
 #include <iterator>  // for iterators
 #include <vector>    // for vectors
-//#include<conio.h>
 #include <math.h>
-//#include <pcl/ros/conversions.h>
 #include <pcl_conversions/pcl_conversions.h>
-//#include <pcl/conversions.h>
 #include <arm_vision/SegmentedClustersArray.h>
 #include <pcl/ModelCoefficients.h>
 #include <pcl/filters/extract_indices.h>
@@ -22,6 +24,7 @@
 #include <pcl/sample_consensus/model_types.h>
 #include <pcl/segmentation/extract_clusters.h>
 #include <pcl/segmentation/sac_segmentation.h>
-//#include <obj_recognition/ClusterData.h>
 
 // PCL specific includes
+
+#endif
