+++ catkin_home/src/arm_vision/include/SegmentationNode/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library(SegmentationNode SegmentationNode.cpp)
